requirements.txt
python-telegram-bot==20.7
playwright==1.47.0
